feat: add EVM indexer
To make indexers extractable it was separated into Indexer class
which instance consumer initiaties. This allows better separation
(no need for pulling deps for all networks) - everything currently
is exported under either evm or starknet object, it can be extracted
later (we might need to extract some common things first into other
package). We also have specific types for writers for each network.

This could also be useful in the future if we put multiple APIs
in single instance of checkpoint, it could accept indexers instead
of just single indexer.
