+++ src/index.ts
@@ -3,6 +3,7 @@ export { LogLevel } from './utils/logger';
 export { AsyncMySqlPool } from './mysql';
 export { createGetLoader } from './graphql';
 export { Model } from './orm';
+export * from './providers';
 export * from './types';
 
 export default Checkpoint;
